[#135] refactor : Image 클래스 VO package로 이동
